requirements.txt
jieba
flask
tweepy
markovify
apscheduler
python-telegram-bot
